requirements.txt
google-cloud-vision==3.4.4
opencv-python==4.8.1.78
numpy==1.24.3
python-dotenv==1.0.0
SQLAlchemy==2.0.23
face-recognition==1.3.0